+++ src/compas_rhino/geometry/surfaces/__init__.py
@@ -12,6 +12,11 @@ def new_surface(cls, *args, **kwargs):
 
 
 @plugin(category="factories", requires=["Rhino"])
+def new_surface_from_plane(cls, *args, **kwargs):
+    return RhinoSurface.from_plane(*args, **kwargs)
+
+
+@plugin(category="factories", requires=["Rhino"])
 def new_nurbssurface(cls, *args, **kwargs):
     return super(NurbsSurface, cls).__new__(cls)